Appointemtn has Finished

+++ app/src/main/java/com/khhhm/worefa/data/repository/AppointmentRepository.kt
@@ -1,13 +1,10 @@
 package com.khhhm.worefa.data.repository
 
 import android.content.Context
-import android.util.Log
 import androidx.lifecycle.LiveData
 import com.khhhm.worefa.data.dao.AppointmentDao
-import com.khhhm.worefa.data.dao.CompanyDao
 import com.khhhm.worefa.data.entitys.Appointment
 import com.khhhm.worefa.data.network.AppointmentApiService
-import com.khhhm.worefa.utilites.isNetworkConnected
 import kotlinx.coroutines.*
 import kotlinx.coroutines.Dispatchers.IO
 
@@ -15,39 +12,41 @@ class AppointmentRepository private constructor(private val appointmentDao: Appo
 
     fun getAllMyAppointment(context: Context):LiveData<List<Appointment>> {
 
-        GlobalScope.launch (IO ) {
-            if(isNetworkConnected(context)){
-                val appointments= appointmentApiService.getMyAppointments("kiber").await().body()
-                if(appointments!=null)     appointmentDao.insertAll(appointments)
-                Log.i("THis is inside Context","Services")
-
-            }
-        }
-        Log.i("THis is out of Context","Services")
         return this.appointmentDao.getAllappointments()
+    }
+    suspend fun getAllFromApi(phoneno:String){
+        withContext(IO){
+            val appointemnts=appointmentApiService.getMyAppointments(phoneno).await().body()
+            if(appointemnts!=null)
+                appointmentDao.insertAll(appointemnts)
+        }
     }
 
     fun getAppointment(appId:Int) = this.appointmentDao.getAppointment(appId)
 
     suspend fun insertAppointment(appointment: Appointment){
         withContext(IO){
-            appointmentDao.insertAppointment(appointment)
+            val newAppointment=
+                appointmentApiService.postAppointments(appointment).await().body()
+           appointmentDao.insertAppointment(newAppointment!!)
         }
     }
     suspend fun deleteAppointment(appointment: Appointment){
         withContext(IO){
-            appointmentDao.deleteAppointment(appointment)
+            if(appointmentApiService.deleteAppointments(appointment.Id).await().isSuccessful){
+                appointmentDao.deleteAppointment(appointment)
+            }
         }
     }
     suspend fun updateAppointment(appointment: Appointment){
         withContext(IO){
-            appointmentDao.updateAppointment(appointment)
+            val updatedAppointment=appointmentApiService.updateAppointments(appointment).await().body()
+            appointmentDao.updateAppointment(updatedAppointment!!)
         }
     }
 
     companion object{
         @Volatile private var instance: AppointmentRepository? = null
-
         fun getInstance(appointmentDao: AppointmentDao,appointmentApiService: AppointmentApiService) =
             instance ?: synchronized(this) {
                 instance ?: AppointmentRepository(appointmentDao,appointmentApiService).also { instance = it }
